ASVO-270 Create record filter section
Moved record fitler section from light-cone into its own form. Refactored
code to use a UI Holder that allows forms to grab values from each other (although
forms have to be created in right order.) Also, added the is_filter to the filters.
The XML is broken and test code was disabled as there is a coming story on this.