Automated Tests: Wait for an element to appear
The current test structure is backwards in that it navigates, waits an arbitrary time and then assumes that the navigation is complete.  This means that a large wait time is required to be confident that the page has appeared.

This patch puts the framework in place to retry finding an element with short waits, allowing the total wait time to be large while not slowing down the tests unacceptably.